fix review comments - merge both import controllers
